requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
openpyxl==3.0.9
psycopg2-binary==2.9.2
numpy
pandas
fasttext